backend/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "adaptive-daltonization-backend"
version = "0.1.0"
description = "Flask service for adaptive color correction using daltonization and TensorFlow Lite."
requires-python = ">=3.10"
dependencies = [
    "Flask>=3.0",
    "numpy>=1.26",
    "scikit-learn>=1.3",
    "scikit-image>=0.22",  # For LAB color space conversion
    "pydantic>=1.10,<2.0",
    "Pillow>=10.0",
    "firebase-admin>=6.2",
    "python-dotenv>=1.0",
    "requests>=2.32",
]

[project.optional-dependencies]
tflite = [
    "tensorflow>=2.16",
]

tests = [
    "pytest>=7.4",
]
